Update dependencies and refactor file structure for improved organization and performance
