+++ frontend/src/App.tsx
@@ -4,8 +4,8 @@ import { Toaster as Sonner } from "@/components/ui/sonner";
 import { TooltipProvider } from "@/components/ui/tooltip";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ThemeProvider } from "next-themes";
-import Index from "@/page/Index";
-import MovieDetails from "@/page/MovieDetails";
+import Index from "@/pages/home/Index";
+import MovieDetails from "@/pages/movie/MovieDetails";
 import NotFound from "@/page/NotFound";
 import LoginForm from "@/pages/auth/LoginPage";
 import About from "./pages/about/About";
