+++ frontend/src/index.css
@@ -15,3 +15,11 @@
 .movie-card-overlay {
   @apply absolute inset-0 bg-black opacity-25;
 }
+.hover-lift {
+  transition: transform 0.3s ease, box-shadow 0.3s ease;
+}
+
+.hover-lift:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+}
