+++ frontend/tailwind.config.ts
@@ -1,5 +1,5 @@
-import type { Config } from "tailwindcss";
 import tailwindcssAnimate from "tailwindcss-animate";
+
 export default {
   darkMode: "class",
   content: [
@@ -125,4 +125,4 @@ export default {
     },
   },
   plugins: [tailwindcssAnimate],
-} satisfies Config;
+};
